requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
cohere==5.5.3
qdrant-client==1.8.0
openai==1.3.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
requests==2.31.0
tiktoken==0.5.2
beautifulsoup4==4.12.2
lxml==4.9.3
asyncio==3.4.3
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2